Add configuration to RSS items to limit the feed to a given date
